requirements.txt
streamlit>=1.28.0
torch>=2.0.1
transformers>=4.35.0
Pillow>=10.2.0
